+++ commands/limit_others.go
@@ -1,3 +1,4 @@
+// +build !darwin
 // Copyright © 2013 Steve Francia <spf@spf13.com>.
 //
 // Licensed under the Simple Public License, Version 2.0 (the "License");
@@ -13,19 +14,6 @@
 
 package commands
 
-import (
-	"github.com/spf13/cobra"
-	"github.com/spf13/hugo/hugolib"
-)
-
-var check = &cobra.Command{
-	Use:   "check",
-	Short: "Check content in the source directory",
-	Long: `Hugo will perform some basic analysis on the
-    content provided and will give feedback.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		InitializeConfig()
-		site := hugolib.Site{Config: *Config}
-		site.Analyze()
-	},
+func tweakLimit() {
+	// nothing to do
 }
